add thymeleaf-layout

+++ src/main/java/com/cho_co_song_i/yummy/yummy/dto/UserProfileDto.java
@@ -2,6 +2,8 @@ package com.cho_co_song_i.yummy.yummy.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
+
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -13,7 +15,7 @@ public class UserProfileDto {
     private String tokenId;
     private String addrType;
     private String addr;
-    private Double lngx;
-    private Double laty;
+    private BigDecimal lngx;
+    private BigDecimal laty;
     private Date regDt;
 }
